correction

+++ Epsilon/Function.h
@@ -5,7 +5,7 @@
 #include <iostream>
 
 using namespace std;
-int pow(double x, int i);
+double pow(double x, int i);
 double g(double x, int i);
 double sum(double x, double epsilon);
 void out(double first, double second,int i);
